Added event parsers

+++ Sadie.Networking.Writers/Rooms/RoomSettingsSavedWriter.cs
@@ -5,9 +5,9 @@ namespace Sadie.Networking.Writers.Rooms;
 
 public class RoomSettingsSavedWriter : NetworkPacketWriter
 {
-    public RoomSettingsSavedWriter(int roomId)
+    public RoomSettingsSavedWriter(long roomId)
     {
         WriteShort(ServerPacketId.RoomSettingsSaved);
-        WriteInteger(roomId);
+        WriteLong(roomId);
     }
 }
