+++ Sadie.Networking.Writers/Rooms/RoomSettingsUpdatedWriter.cs
@@ -5,9 +5,9 @@ namespace Sadie.Networking.Writers.Rooms;
 
 public class RoomSettingsUpdatedWriter : NetworkPacketWriter
 {
-    public RoomSettingsUpdatedWriter(int roomId)
+    public RoomSettingsUpdatedWriter(long roomId)
     {
         WriteShort(ServerPacketId.RoomSettingsUpdated);
-        WriteInteger(roomId);
+        WriteLong(roomId);
     }
 }
